FlaskRockPaperScissorsGame/rest_api_game/requirements.txt
flask-restx==1.0.5
Flask-SQLAlchemy==2.5.1
Werkzeug==2.0.1
